update patch for nvfuser installation
